finish ft_isascii.c & ft_isprint.c + put main functions in separate files

+++ ft_strlcpy.c
@@ -1,6 +1,6 @@
-size_t	ft_strlcpy(char * dst, const char *src, size_t dstsize)
+int	ft_strlcpy(char *dst, const char *src, int dstsize)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	if (dstsize > 0)
